requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
twilio==8.9.1
supabase==2.16.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
phonenumbers==8.13.24
APScheduler==3.10.4 